# docs/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Casa da Paz API
  version: 0.1.0
servers:
  - url: /api
paths:
...
  /marketing/resumo:
    get:
      security: [{ bearerAuth: [] }]
      summary: KPIs publicação
  /marketing/eventos:
    get:
      security: [{ bearerAuth: [] }]
    post:
      security: [{ bearerAuth: [] }]
      summary: Publicar evento
  /marketing/produtos:
    get:
      security: [{ bearerAuth: [] }]
  /eventos:
    get:
      security: [{ bearerAuth: [] }]
  /eventos/{id}/inscricoes/{inscricaoId}/cobrar:
    post:
      security: [{ bearerAuth: [] }]
      summary: Cobrança Asaas para inscrição
  /eventos/{id}/checkin:
    post:
      security: [{ bearerAuth: [] }]
  /livraria/produtos:
    get:
      security: [{ bearerAuth: [] }]
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT